Devtool is a component, move it to component dir
